refactor(🧹): Clean up AREnableArtworksConnectionForAuction ff
