+++ src/app/Components/ArtworkFilter/ArtworkFilterHelpers.ts
@@ -83,9 +83,7 @@ export enum ViewAsValues {
 export const getSortDefaultValueByFilterType = (filterType: FilterType) => {
   return {
     artwork: "-decayed_merch",
-    saleArtwork: "position",
-    // TODO: Replace newSaleArtwork with saleArtwork when AREnableArtworksConnectionForAuction is released
-    newSaleArtwork: "sale_position",
+    saleArtwork: "sale_position",
     showArtwork: "partner_show_position",
     auctionResult: "DATE_DESC",
     geneArtwork: "-partner_updated_at",
@@ -205,8 +203,6 @@ export type FilterArray = ReadonlyArray<FilterData>
 export type FilterType =
   | "artwork"
   | "saleArtwork"
-  // TODO: Replace newSaleArtwork with saleArtwork when AREnableArtworksConnectionForAuction is released
-  | "newSaleArtwork"
   | "showArtwork"
   | "auctionResult"
   | "geneArtwork"
@@ -257,12 +253,6 @@ const DEFAULT_ARTWORKS_PARAMS = {
 } as FilterParams
 
 const DEFAULT_SALE_ARTWORKS_PARAMS = {
-  sort: "position",
-  estimateRange: "",
-} as FilterParams
-
-// TODO: Replace DEFAULT_NEW_SALE_ARTWORKS_PARAMS with DEFAULT_SALE_ARTWORKS_PARAMS when AREnableArtworksConnectionForAuction is released
-const DEFAULT_NEW_SALE_ARTWORKS_PARAMS = {
   sort: "sale_position",
   estimateRange: "",
 } as FilterParams
@@ -306,8 +296,6 @@ const getDefaultParamsByType = (filterType: FilterType) => {
   return {
     artwork: DEFAULT_ARTWORKS_PARAMS,
     saleArtwork: DEFAULT_SALE_ARTWORKS_PARAMS,
-    // TODO: Replace newSaleArtwork with saleArtwork when AREnableArtworksConnectionForAuction is released
-    newSaleArtwork: DEFAULT_NEW_SALE_ARTWORKS_PARAMS,
     showArtwork: DEFAULT_SHOW_ARTWORKS_PARAMS,
     auctionResult: DEFAULT_AUCTION_RESULT_PARAMS,
     geneArtwork: DEFAULT_GENE_ARTWORK_PARAMS,
