+++ src/app/Components/ArtworkFilter/ArtworkFilterStore.tsx
@@ -237,12 +237,6 @@ export const selectedOptionsUnion = ({
     },
     saleArtwork: {
       paramName: FilterParamName.sort,
-      paramValue: "position",
-      displayText: "Lot Number Ascending",
-    },
-    // TODO: Replace newSaleArtwork with saleArtwork when AREnableArtworksConnectionForAuction is released
-    newSaleArtwork: {
-      paramName: FilterParamName.sort,
       paramValue: "sale_position",
       displayText: "Lot Number Ascending",
     },
