+++ src/app/Components/ArtworkFilter/Filters/SortOptions.tsx
@@ -37,7 +37,7 @@ const DEFAULT_TAG_SORT = {
   paramValue: "-partner_updated_at",
 }
 
-export const DEFAULT_NEW_SALE_ARTWORK_SORT = {
+export const DEFAULT_SALE_ARTWORK_SORT = {
   displayText: "Lot Number Ascending",
   paramName: FilterParamName.sort,
   paramValue: "sale_position",
@@ -77,41 +77,7 @@ export const ORDERED_ARTWORK_SORTS: FilterData[] = [
 ]
 
 export const ORDERED_SALE_ARTWORK_SORTS: FilterData[] = [
-  {
-    displayText: "Lot Number Ascending",
-    paramName: FilterParamName.sort,
-    paramValue: "position",
-  },
-  {
-    displayText: "Lot Number Descending",
-    paramName: FilterParamName.sort,
-    paramValue: "-position",
-  },
-  {
-    displayText: "Most Bids",
-    paramName: FilterParamName.sort,
-    paramValue: "-bidder_positions_count",
-  },
-  {
-    displayText: "Least Bids",
-    paramName: FilterParamName.sort,
-    paramValue: "bidder_positions_count",
-  },
-  {
-    displayText: "Highest Bid",
-    paramName: FilterParamName.sort,
-    paramValue: "-searchable_estimate",
-  },
-  {
-    displayText: "Lowest Bid",
-    paramName: FilterParamName.sort,
-    paramValue: "searchable_estimate",
-  },
-]
-
-// TODO: Replace DEFAULT_NEW_SALE_ARTWORKS_PARAMS with DEFAULT_SALE_ARTWORKS_PARAMS when AREnableArtworksConnectionForAuction is released
-export const ORDERED_NEW_SALE_ARTWORK_SORTS: FilterData[] = [
-  DEFAULT_NEW_SALE_ARTWORK_SORT,
+  DEFAULT_SALE_ARTWORK_SORT,
   {
     displayText: "Lot Number Descending",
     paramName: FilterParamName.sort,
@@ -172,8 +138,6 @@ export const SortOptionsScreen: React.FC<SortOptionsScreenProps> = ({ navigation
   const filterOptions = {
     artwork: [DEFAULT_ARTWORK_SORT, ...ORDERED_ARTWORK_SORTS],
     saleArtwork: ORDERED_SALE_ARTWORK_SORTS,
-    // TODO: Replace newSaleArtwork with saleArtwork when AREnableArtworksConnectionForAuction is released
-    newSaleArtwork: ORDERED_NEW_SALE_ARTWORK_SORTS,
     showArtwork: [GALLERY_CURATED_ARTWORK_SORT, ...ORDERED_ARTWORK_SORTS],
     auctionResult: ORDERED_AUCTION_RESULTS_SORTS,
     geneArtwork: [DEFAULT_GENE_SORT, ...ORDERED_ARTWORK_SORTS],
