+++ src/app/Scenes/Sale/Components/BuyNowArtworksRail.tests.tsx
@@ -1,6 +1,7 @@
+import { screen } from "@testing-library/react-native"
 import { setupTestWrapper } from "app/utils/tests/setupTestWrapper"
 import { graphql } from "react-relay"
-import { BuyNowArtworksRailContainer } from "./Components/BuyNowArtworksRail"
+import { BuyNowArtworksRailContainer } from "./BuyNowArtworksRail"
 
 describe("BuyNowArtworksRail", () => {
   const { renderWithRelay } = setupTestWrapper({
@@ -15,47 +16,47 @@ describe("BuyNowArtworksRail", () => {
     variables: { id: "sale-id" },
   })
 
-  it(`renders title "Buy now"`, () => {
-    const { getByText } = renderWithRelay(mockProps)
-    expect(getByText("Artworks Available to Buy Now")).toBeDefined()
+  it(`renders "Buy now" rail and artworks`, () => {
+    renderWithRelay(mockProps)
+    expect(screen.queryByText("Artworks Available to Buy Now")).toBeDefined()
+    expect(screen.queryAllByText("Best artwork ever, 2019")).toBeDefined()
   })
 
   it("renders nothing if there are no artworks", () => {
     const noArtworksProps = {
       Sale: () => ({
-        saleArtworksConnection: {
+        artworksConnection: {
           edges: [],
         },
       }),
     }
+
     const { queryAllByTestId } = renderWithRelay(noArtworksProps)
     expect(queryAllByTestId("bnmo-rail-wrapper")).toHaveLength(0)
   })
 })
 
-const saleArtworksConnectionEdges = new Array(10).fill({
+const artworksConnectionEdges = new Array(10).fill({
   node: {
-    artwork: {
-      id: "artwork-id",
-      title: "Best artwork ever",
-      date: "2019",
-      saleMessage: "Contact For Price",
-      artistNames: "Banksy",
-      href: "/artwork/artwroks-href",
-      image: {
-        url: "artworkImageUrl",
-      },
-      partner: {
-        name: "Heritage Auctions",
-      },
+    id: "artwork-id",
+    title: "Best artwork ever",
+    date: "2019",
+    saleMessage: "Contact For Price",
+    artistNames: "Banksy",
+    href: "/artwork/artwroks-href",
+    image: {
+      url: "artworkImageUrl",
+    },
+    partner: {
+      name: "Heritage Auctions",
     },
   },
 })
 
 const mockProps = {
   Sale: () => ({
-    saleArtworksConnection: {
-      edges: saleArtworksConnectionEdges,
+    artworksConnection: {
+      edges: artworksConnectionEdges,
     },
   }),
 }
