+++ src/app/Scenes/Sale/Sale.tsx
@@ -2,8 +2,10 @@ import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { Spacer, Flex, Box, Join } from "@artsy/palette-mobile"
 import { captureMessage } from "@sentry/react-native"
 import { SaleAboveTheFoldQuery } from "__generated__/SaleAboveTheFoldQuery.graphql"
-import { SaleBelowTheFoldNewQuery$data } from "__generated__/SaleBelowTheFoldNewQuery.graphql"
-import { SaleBelowTheFoldQuery } from "__generated__/SaleBelowTheFoldQuery.graphql"
+import {
+  SaleBelowTheFoldQuery,
+  SaleBelowTheFoldQuery$data,
+} from "__generated__/SaleBelowTheFoldQuery.graphql"
 import { Sale_me$data } from "__generated__/Sale_me.graphql"
 import { Sale_sale$data } from "__generated__/Sale_sale.graphql"
 import {
@@ -12,16 +14,16 @@ import {
   FilterModalMode,
 } from "app/Components/ArtworkFilter"
 import { ArtworkFiltersStoreProvider } from "app/Components/ArtworkFilter/ArtworkFilterStore"
-import { DEFAULT_NEW_SALE_ARTWORK_SORT } from "app/Components/ArtworkFilter/Filters/SortOptions"
+import { DEFAULT_SALE_ARTWORK_SORT } from "app/Components/ArtworkFilter/Filters/SortOptions"
 import { LoadFailureView } from "app/Components/LoadFailureView"
 import Spinner from "app/Components/Spinner"
 import { CascadingEndTimesBanner } from "app/Scenes/Artwork/Components/CascadingEndTimesBanner"
+import { SaleLotsListContainer } from "app/Scenes/Sale/Components/SaleLotsList"
 import { unsafe__getEnvironment } from "app/store/GlobalStore"
 import { navigate, popParentViewController } from "app/system/navigation/navigate"
 import { getRelayEnvironment } from "app/system/relay/defaultEnvironment"
 import { AboveTheFoldQueryRenderer } from "app/utils/AboveTheFoldQueryRenderer"
 import { AuctionWebsocketContextProvider } from "app/utils/Websockets/auctions/AuctionSocketContext"
-import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
 import { PlaceholderBox, PlaceholderText, ProvidePlaceholderContext } from "app/utils/placeholders"
 import { ProvideScreenTracking, Schema } from "app/utils/track"
 import _, { times } from "lodash"
@@ -34,13 +36,10 @@ import useInterval from "react-use/lib/useInterval"
 import usePrevious from "react-use/lib/usePrevious"
 import RelayModernEnvironment from "relay-runtime/lib/store/RelayModernEnvironment"
 import { BuyNowArtworksRailContainer } from "./Components/BuyNowArtworksRail"
-import { NewBuyNowArtworksRailContainer } from "./Components/NewBuyNowArtworksRail"
-import { NewSaleLotsListContainer } from "./Components/NewSaleLotsList"
 import { RegisterToBidButtonContainer } from "./Components/RegisterToBidButton"
 import { SaleActiveBidsContainer } from "./Components/SaleActiveBids"
 import { SaleArtworksRailContainer } from "./Components/SaleArtworksRail"
 import { COVER_IMAGE_HEIGHT, SaleHeaderContainer as SaleHeader } from "./Components/SaleHeader"
-import { SaleLotsListContainer } from "./Components/SaleLotsList"
 import { saleStatus } from "./helpers"
 
 interface Props {
@@ -80,7 +79,6 @@ const NOOP = () => {}
 
 export const Sale: React.FC<Props> = ({ sale, me, below, relay }) => {
   const tracking = useTracking()
-  const enableArtworksConnection = useFeatureFlag("AREnableArtworksConnectionForAuction")
 
   const flatListRef = useRef<FlatList<any>>(null)
   const [isRefreshing, setIsRefreshing] = useState(false)
@@ -169,25 +167,12 @@ export const Sale: React.FC<Props> = ({ sale, me, below, relay }) => {
 
   const renderSaleLotsList = () => {
     if (below) {
-      if (enableArtworksConnection) {
-        return (
-          <NewSaleLotsListContainer
-            unfilteredArtworks={
-              (below as unknown as SaleBelowTheFoldNewQuery$data).viewer?.unfilteredArtworks!
-            }
-            viewer={(below as unknown as SaleBelowTheFoldNewQuery$data).viewer}
-            saleID={sale.internalID}
-            saleSlug={sale.slug}
-            scrollToTop={scrollToTop}
-            artworksRefetchRef={artworksRefetchRef}
-          />
-        )
-      }
-
       return (
         <SaleLotsListContainer
-          saleArtworksConnection={below}
-          unfilteredSaleArtworksConnection={below.unfilteredSaleArtworksConnection}
+          unfilteredArtworks={
+            (below as unknown as SaleBelowTheFoldQuery$data).viewer?.unfilteredArtworks!
+          }
+          viewer={(below as unknown as SaleBelowTheFoldQuery$data).viewer}
           saleID={sale.internalID}
           saleSlug={sale.slug}
           scrollToTop={scrollToTop}
@@ -245,11 +230,7 @@ export const Sale: React.FC<Props> = ({ sale, me, below, relay }) => {
     },
     {
       key: BUY_NOW_ARTWORKS_RAIL,
-      content: enableArtworksConnection ? (
-        <NewBuyNowArtworksRailContainer sale={sale} />
-      ) : (
-        <BuyNowArtworksRailContainer sale={sale} />
-      ),
+      content: <BuyNowArtworksRailContainer sale={sale} />,
     },
     {
       key: SALE_LOTS_LIST,
@@ -392,7 +373,6 @@ export const SaleContainer = createRefetchContainer(
         ...SaleHeader_sale
         ...RegisterToBidButton_sale
         ...BuyNowArtworksRail_sale
-        ...NewBuyNowArtworksRail_sale
         endAt
         internalID
         liveStartAt
@@ -429,31 +409,16 @@ export const SaleScreenQuery = graphql`
 `
 
 const SaleScreenBelowQuery = graphql`
-  query SaleBelowTheFoldQuery($saleID: ID) {
-    ...SaleLotsList_saleArtworksConnection @arguments(saleID: $saleID)
-    unfilteredSaleArtworksConnection: saleArtworksConnection(
-      saleID: $saleID
-      aggregations: [TOTAL]
-    ) {
-      ...SaleLotsList_unfilteredSaleArtworksConnection
-      counts {
-        total
-      }
-    }
-  }
-`
-
-const SaleScreenBelowNewQuery = graphql`
-  query SaleBelowTheFoldNewQuery($saleID: ID, $input: FilterArtworksInput) {
+  query SaleBelowTheFoldQuery($saleID: ID, $input: FilterArtworksInput) {
     viewer {
       unfilteredArtworks: artworksConnection(
         saleID: $saleID
         aggregations: [FOLLOWED_ARTISTS, ARTIST, MEDIUM, TOTAL]
         first: 0
       ) {
-        ...NewSaleLotsList_unfilteredArtworks
+        ...SaleLotsList_unfilteredArtworks
       }
-      ...NewSaleLotsList_viewer @arguments(saleID: $saleID, input: $input)
+      ...SaleLotsList_viewer @arguments(saleID: $saleID, input: $input)
     }
   }
 `
@@ -463,7 +428,6 @@ export const SaleQueryRenderer: React.FC<{
   environment?: RelayModernEnvironment
 }> = ({ saleID, environment }) => {
   const { trackEvent } = useTracking()
-  const enableArtworksConnection = useFeatureFlag("AREnableArtworksConnectionForAuction")
 
   useEffect(() => {
     trackEvent(tracks.pageView(saleID))
@@ -476,24 +440,17 @@ export const SaleQueryRenderer: React.FC<{
         query: SaleScreenQuery,
         variables: { saleID, saleSlug: saleID },
       }}
-      below={
-        enableArtworksConnection
-          ? {
-              query: SaleScreenBelowNewQuery,
-              variables: {
-                saleID,
-                // @ts-ignore
-                input: {
-                  sort: DEFAULT_NEW_SALE_ARTWORK_SORT.paramValue,
-                  priceRange: "",
-                },
-              },
-            }
-          : {
-              query: SaleScreenBelowQuery,
-              variables: { saleID },
-            }
-      }
+      below={{
+        query: SaleScreenBelowQuery,
+        variables: {
+          saleID,
+          // @ts-ignore
+          input: {
+            sort: DEFAULT_SALE_ARTWORK_SORT.paramValue,
+            priceRange: "",
+          },
+        },
+      }}
       render={({ props, error }) => {
         if (error) {
           if (__DEV__) {
